+++ src/pages/programsWip/form/ProgramNotificationsFormContents.tsx
@@ -1,7 +1,9 @@
 import i18n from '@dhis2/d2-i18n'
 import { Button, IconAdd16, NoticeBox } from '@dhis2/ui'
-import React from 'react'
+import React, { Dispatch, SetStateAction } from 'react'
+import { useFormState } from 'react-final-form'
 import { useFieldArray } from 'react-final-form-arrays'
+import { useParams } from 'react-router-dom'
 import {
     DrawerPortal,
     SectionedFormSection,
@@ -16,6 +18,7 @@ import {
     SubmittedNotificationFormValues,
 } from './programNotification/NotificationForm'
 import css from './ProgramStagesForm.module.css'
+import { ProgramStageListItem } from './ProgramStagesFormContents'
 
 export type ProgramNotificationListItem = {
     id: string
@@ -24,6 +27,7 @@ export type ProgramNotificationListItem = {
     access?: Access
     program?: { id: string }
 }
+type NotificationFormOpen = DisplayableModel | null | undefined
 
 export const ProgramNotificationsFormContents = React.memo(
     function ProgramNotificationsContents({ name }: { name: string }) {
@@ -41,14 +45,42 @@ export const ProgramNotificationsFormContents = React.memo(
     }
 )
 
-const NotificationListNewOrEdit = () => {
+const DeletedItem = ({
+    id,
+    displayName,
+    index,
+    handleCancelDelete,
+}: {
+    id: string
+    displayName: string
+    index: number
+    handleCancelDelete: (index: number) => void
+}) => (
+    <div className={css.stageCardDeleted} key={id}>
+        <div className={css.deletedStageText}>
+            {i18n.t(
+                'Notification {{- notificationName}} will be removed on save',
+                { notificationName: displayName }
+            )}
+        </div>
+
+        <Button small onClick={() => handleCancelDelete(index)}>
+            {i18n.t('Restore notification')}
+        </Button>
+    </div>
+)
+
+const ProgramNotificationListNewOrEdit = ({
+    setNotificationFormOpen,
+    notificationFormOpen,
+}: {
+    setNotificationFormOpen: Dispatch<SetStateAction<NotificationFormOpen>>
+    notificationFormOpen: NotificationFormOpen
+}) => {
     const programNotificationsFieldArray =
         useFieldArray<ProgramNotificationListItem>(
             'notificationTemplates'
         ).fields
-    const [notificationFormOpen, setNotificationFormOpen] = React.useState<
-        DisplayableModel | null | undefined
-    >()
     const isNotificationFormOpen =
         !!notificationFormOpen || notificationFormOpen === null
 
@@ -79,6 +111,20 @@ const NotificationListNewOrEdit = () => {
         }
     }
 
+    const handleDeletedProgramNotification = (index: number) => {
+        programNotificationsFieldArray.update(index, {
+            ...programNotificationsFieldArray.value[index],
+            deleted: true,
+        })
+    }
+
+    const handleCancelDeletedProgramNotification = (index: number) => {
+        programNotificationsFieldArray.update(index, {
+            ...programNotificationsFieldArray.value[index],
+            deleted: false,
+        })
+    }
+
     const onCloseNotificationForm = () => {
         setNotificationFormOpen(undefined)
     }
@@ -95,6 +141,9 @@ const NotificationListNewOrEdit = () => {
                             notification={notificationFormOpen}
                             onCancel={onCloseNotificationForm}
                             onSubmitted={handleSubmittedNotification}
+                            notificationList={programNotificationsFieldArray.value.map(
+                                (n) => ({ id: n.id })
+                            )}
                         />
                     </div>
                 )}
@@ -103,30 +152,232 @@ const NotificationListNewOrEdit = () => {
             <div className={css.listWrapper}>
                 {(programNotificationsFieldArray.value ?? []).length === 0 && (
                     <NoticeBox className={css.formTypeInfo}>
-                        {i18n.t('No notifications been added yet')}
+                        {i18n.t('No notifications have been added yet')}
                     </NoticeBox>
                 )}
-
-                <div className={css.sectionItems}>
-                    {(programNotificationsFieldArray.value ?? []).map(
-                        (notification) => {
+                {(programNotificationsFieldArray.value ?? []).map(
+                    (notification, index) => {
+                        if (notification.deleted) {
                             return (
-                                <ListInFormItem
+                                <DeletedItem
                                     key={notification.id}
-                                    item={notification}
-                                    schemaName={
-                                        SchemaName.programNotificationTemplate
+                                    id={notification.id}
+                                    displayName={notification.displayName}
+                                    index={index}
+                                    handleCancelDelete={
+                                        handleCancelDeletedProgramNotification
                                     }
-                                    onClick={() =>
-                                        setNotificationFormOpen(notification)
-                                    }
-                                    onDelete={() => {}}
                                 />
                             )
                         }
-                    )}
-                </div>
+                        return (
+                            <ListInFormItem
+                                key={notification.id}
+                                item={notification}
+                                schemaName={
+                                    SchemaName.programNotificationTemplate
+                                }
+                                onClick={() =>
+                                    setNotificationFormOpen(notification)
+                                }
+                                onDelete={() =>
+                                    handleDeletedProgramNotification(index)
+                                }
+                            />
+                        )
+                    }
+                )}
+            </div>
+        </>
+    )
+}
 
+const StageNotificationListNewOrEdit = ({
+    stage,
+    stageIndex,
+    setNotificationFormOpen,
+    notificationFormOpen,
+}: {
+    stage: ProgramStageListItem
+    stageIndex: number
+    setNotificationFormOpen: Dispatch<SetStateAction<NotificationFormOpen>>
+    notificationFormOpen: NotificationFormOpen
+}) => {
+    const stagesFieldArray =
+        useFieldArray<ProgramStageListItem>('programStages').fields
+    const isNotificationFormOpen =
+        !!notificationFormOpen || notificationFormOpen === null
+    const notificationsArray =
+        stagesFieldArray?.value?.[stageIndex]?.notificationTemplates || []
+
+    const handleSubmittedNotification = (
+        values: SubmittedNotificationFormValues,
+        closeOnSubmit: boolean = true
+    ) => {
+        const isEditNotification =
+            notificationFormOpen && notificationFormOpen.id
+
+        if (closeOnSubmit) {
+            setNotificationFormOpen(undefined)
+        } else if (!isEditNotification) {
+            setNotificationFormOpen({
+                id: values.id,
+                displayName: values.displayName,
+            })
+        }
+        if (isEditNotification) {
+            const index = notificationsArray.findIndex(
+                (s) => s.id === notificationFormOpen.id
+            )
+            if (index !== -1) {
+                const updatedStageNotifications = [
+                    ...notificationsArray.slice(0, index),
+                    { ...values, name: values?.name ?? values.displayName },
+                    ...notificationsArray.slice(index + 1),
+                ]
+                stagesFieldArray.update(stageIndex, {
+                    ...stage,
+                    notificationTemplates: updatedStageNotifications,
+                })
+            }
+        } else {
+            const updatedStageNotifications = [
+                ...notificationsArray,
+                { ...values, name: values?.name ?? values.displayName },
+            ]
+            stagesFieldArray.update(stageIndex, {
+                ...stage,
+                notificationTemplates: updatedStageNotifications,
+            })
+        }
+    }
+
+    const handleDeletedProgramStageNotification = (index: number) => {
+        const updatedStageNotifications = [
+            ...notificationsArray.slice(0, index),
+            { ...notificationsArray[index], deleted: true },
+            ...notificationsArray.slice(index + 1),
+        ]
+        stagesFieldArray.update(stageIndex, {
+            ...stage,
+            notificationTemplates: updatedStageNotifications,
+        })
+    }
+
+    const handleCancelDeletedProgramStageNotification = (index: number) => {
+        const updatedStageNotifications = [
+            ...notificationsArray.slice(0, index),
+            { ...notificationsArray[index], deleted: false },
+            ...notificationsArray.slice(index + 1),
+        ]
+        stagesFieldArray.update(stageIndex, {
+            ...stage,
+            notificationTemplates: updatedStageNotifications,
+        })
+    }
+
+    const onCloseNotificationForm = () => {
+        setNotificationFormOpen(undefined)
+    }
+
+    return (
+        <>
+            <DrawerPortal
+                isOpen={isNotificationFormOpen}
+                onClose={onCloseNotificationForm}
+            >
+                {notificationFormOpen !== undefined && (
+                    <div>
+                        <EditOrNewNotificationForm
+                            notification={notificationFormOpen}
+                            onCancel={onCloseNotificationForm}
+                            onSubmitted={handleSubmittedNotification}
+                            notificationList={notificationsArray.map((n) => ({
+                                id: n.id,
+                            }))}
+                        />
+                    </div>
+                )}
+            </DrawerPortal>
+
+            {notificationsArray.map((notification, index) => {
+                if (notification.deleted) {
+                    return (
+                        <DeletedItem
+                            key={notification.id}
+                            index={index}
+                            id={notification.id}
+                            displayName={notification.displayName}
+                            handleCancelDelete={
+                                handleCancelDeletedProgramStageNotification
+                            }
+                        />
+                    )
+                }
+                return (
+                    <ListInFormItem
+                        key={notification.id}
+                        item={notification}
+                        schemaName={SchemaName.programNotificationTemplate}
+                        onClick={() => setNotificationFormOpen(notification)}
+                        onDelete={() => {
+                            handleDeletedProgramStageNotification(index)
+                        }}
+                    />
+                )
+            })}
+        </>
+    )
+}
+
+const NotificationListNewOrEdit = () => {
+    const modelId = useParams().id as string
+    const { values } = useFormState({ subscription: { values: true } })
+    // TODO: might want to show the to be deleted notification with a warning instead
+    const stages: ProgramStageListItem[] =
+        values.programStages?.filter(
+            (stage: ProgramStageListItem) => !stage.deleted
+        ) || []
+
+    const [notificationFormOpen, setNotificationFormOpen] =
+        React.useState<NotificationFormOpen>()
+
+    if (!modelId) {
+        return (
+            <NoticeBox className={css.formTypeInfo}>
+                {i18n.t(
+                    'Program must be saved before notifications can be added'
+                )}
+            </NoticeBox>
+        )
+    }
+
+    return (
+        <>
+            <div className={css.listWrapper}>
+                <div className={css.sectionItems}>
+                    {stages.every(
+                        (s) => s?.notificationTemplates?.length === 0
+                    ) &&
+                        values?.notificationTemplates?.length === 0 && (
+                            <NoticeBox className={css.formTypeInfo}>
+                                {i18n.t('No notifications have been added yet')}
+                            </NoticeBox>
+                        )}
+                    <ProgramNotificationListNewOrEdit
+                        setNotificationFormOpen={setNotificationFormOpen}
+                        notificationFormOpen={notificationFormOpen}
+                    />
+                    {stages.map((stage, index) => (
+                        <StageNotificationListNewOrEdit
+                            stage={stage}
+                            key={stage.id}
+                            setNotificationFormOpen={setNotificationFormOpen}
+                            notificationFormOpen={notificationFormOpen}
+                            stageIndex={index}
+                        />
+                    ))}
+                </div>
                 <div>
                     <Button
                         secondary